// packages/frontend/src/types/schemas/prisma-generated/inputTypeSchemas/UserWhereUniqueInputSchema.ts
export const UserWhereUniqueInputSchema: z.ZodType<Prisma.UserWhereUniqueInput> = z.union([
  z.object({
    id: z.string().uuid(),
    email: z.string(),
    firebaseUid: z.string(),
    verificationToken: z.string()
  }),
  z.object({
    id: z.string().uuid(),
    email: z.string(),
    firebaseUid: z.string(),
  }),
  z.object({
    id: z.string().uuid(),
    email: z.string(),
    verificationToken: z.string(),
  }),
  z.object({
    id: z.string().uuid(),
    email: z.string(),
  }),
  z.object({
    id: z.string().uuid(),
    firebaseUid: z.string(),
    verificationToken: z.string(),
  }),
  z.object({
    id: z.string().uuid(),
    firebaseUid: z.string(),
  }),
  z.object({
    id: z.string().uuid(),
    verificationToken: z.string(),
  }),
  z.object({
    id: z.string().uuid(),
  }),
  z.object({
    email: z.string(),
    firebaseUid: z.string(),
    verificationToken: z.string(),
  }),
  z.object({
    email: z.string(),
    firebaseUid: z.string(),
  }),
  z.object({
    email: z.string(),
    verificationToken: z.string(),
  }),
  z.object({
    email: z.string(),
  }),
  z.object({
    firebaseUid: z.string(),
    verificationToken: z.string(),
  }),
  z.object({
    firebaseUid: z.string(),
  }),
  z.object({
    verificationToken: z.string(),
  }),
])
.and(z.object({
  id: z.string().uuid().optional(),
  email: z.string().optional(),
  firebaseUid: z.string().optional(),
  verificationToken: z.string().optional(),
  AND: z.union([ z.lazy(() => UserWhereInputSchema),z.lazy(() => UserWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => UserWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => UserWhereInputSchema),z.lazy(() => UserWhereInputSchema).array() ]).optional(),
  name: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  role: z.union([ z.lazy(() => EnumRoleFilterSchema),z.lazy(() => RoleSchema) ]).optional(),
  planType: z.union([ z.lazy(() => EnumPlanTypeFilterSchema),z.lazy(() => PlanTypeSchema) ]).optional(),
  isAdmin: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  isVerified: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  resetPasswordToken: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  resetPasswordExpires: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  invitationExpires: z.union([ z.lazy(() => DateTimeNullableFilterSchema),z.coerce.date() ]).optional().nullable(),
  invitedBy: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  updatedAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  isFirstLogin: z.union([ z.lazy(() => BoolFilterSchema),z.boolean() ]).optional(),
  profile: z.union([ z.lazy(() => ProfileNullableScalarRelationFilterSchema),z.lazy(() => ProfileWhereInputSchema) ]).optional().nullable(),
  projects: z.lazy(() => ProjectListRelationFilterSchema).optional(),
  sentMessages: z.lazy(() => MessageListRelationFilterSchema).optional(),
  receivedMessages: z.lazy(() => MessageListRelationFilterSchema).optional(),
  bookmarks: z.lazy(() => BookmarkListRelationFilterSchema).optional(),
  reports: z.lazy(() => ReportListRelationFilterSchema).optional(),
  notifications: z.lazy(() => NotificationListRelationFilterSchema).optional(),
  notificationArchives: z.lazy(() => NotificationArchiveListRelationFilterSchema).optional(),
  notificationSetting: z.union([ z.lazy(() => NotificationSettingNullableScalarRelationFilterSchema),z.lazy(() => NotificationSettingWhereInputSchema) ]).optional().nullable(),
  subscription: z.union([ z.lazy(() => SubscriptionNullableScalarRelationFilterSchema),z.lazy(() => SubscriptionWhereInputSchema) ]).optional().nullable(),
  pendingSubscription: z.union([ z.lazy(() => PendingSubscriptionNullableScalarRelationFilterSchema),z.lazy(() => PendingSubscriptionWhereInputSchema) ]).optional().nullable(),
  subscriptionHistory: z.lazy(() => SubscriptionHistoryListRelationFilterSchema).optional(),
  activityLogs: z.lazy(() => ActivityLogListRelationFilterSchema).optional(),
  payments: z.lazy(() => PaymentListRelationFilterSchema).optional(),
  errorLogs: z.lazy(() => ErrorLogListRelationFilterSchema).optional(),
  mentionedInMessages: z.lazy(() => MessageListRelationFilterSchema).optional(),
  createdInvitations: z.lazy(() => InvitationCodeListRelationFilterSchema).optional(),
  usedInvitations: z.lazy(() => InvitationCodeListRelationFilterSchema).optional(),
  createdTemplates: z.lazy(() => EmailTemplateListRelationFilterSchema).optional(),
  updatedTemplates: z.lazy(() => EmailTemplateListRelationFilterSchema).optional(),
  sentEmails: z.lazy(() => EmailLogListRelationFilterSchema).optional()
}).strict());